moved profiles to src
